+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - `Config` doesn't force OpenGL `Api` by default.
 - `Display::create_context` now uses the most recent available `Api` from the `Config` when `ContextApi` is not specified in `ContextAttributes`.
 - `PossiblyCurrentGlContext::get_proc_address` method was moved to `GlDisplay::get_proc_address`.
+- `ConfigTemplateBuilder::with_sample_buffers` now called `ConfigTemplateBuilder::with_multisampling`.
+- `GlConfig::sample_buffers` now called `GlConfig::num_samples` and returns the amount of samples in multisample buffer.
 
 # Version 0.30.0-beta.2 (2022-09-03)